feat: add promotion banner component
